+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Unreleased]
+### Fixed
+
+- [TD-4799] Performance issue filtering executions by source
+
 ## [4.43.1] 2022-05-11
 
 - [TD-4089] Support for Implementations without Rules